Feature/trending page sprint2

+++ src/modules/viewinghistory/viewinghistory.service.ts
@@ -5,6 +5,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { ViewingHistory } from './schemas/viewinghistory.entity';
 import aqp from 'api-query-params';
+import { ClearOneViewingHistoryDto } from './dto/clear-one-viewing-history.dto';
+import { ClearAllViewingHistoryDto } from './dto/clear-all-viewing-history.dto';
 
 @Injectable()
 export class ViewinghistoryService {
@@ -14,10 +16,19 @@ export class ViewinghistoryService {
   ) {}
 
   async create(createViewinghistoryDto: CreateViewinghistoryDto) {
-    const { userId } = createViewinghistoryDto;
+    const { userId, videoId } = createViewinghistoryDto;
+
+    const existingHistory = await this.viewingHistoryModel.findOneAndUpdate(
+      { userId, videoId },
+      {},
+      { new: true },
+    );
+
+    if (existingHistory) {
+      return existingHistory;
+    }
 
     const count = await this.viewingHistoryModel.countDocuments({ userId });
-
     if (count >= 100) {
       await this.viewingHistoryModel
         .findOneAndDelete({ userId }, { sort: { createdAt: 1 } })
@@ -32,41 +43,63 @@ export class ViewinghistoryService {
     query: string,
     current: number,
     pageSize: number,
+    searchValue: string,
   ) {
     const { filter, sort } = aqp(query);
-    // Thêm điều kiện lọc theo userId
+
     if (filter.current) delete filter.current;
     if (filter.pageSize) delete filter.pageSize;
+    if (filter.searchValue) delete filter.searchValue;
 
-    if (filter.query) {
-      filter.query = JSON.parse(filter.query); // Chuyển chuỗi JSON thành đối tượng
+    if (filter.query.$in) {
+      filter.query = filter.query.$in;
+    }
+    filter.query = JSON.parse(filter.query);
+
+    if (!filter.query.createdAt || filter.query.createdAt?.length === 0) {
+      filter.query = { userId: filter.query.userId };
+    }
+
+    if (filter.query?.createdAt && typeof filter.query.createdAt === 'string') {
+      const date = new Date(filter.query.createdAt);
+      if (!isNaN(date.getTime())) {
+        const isoDateStr = date.toISOString().split('T')[0];
+
+        filter.query.updatedAt = {
+          $gte: new Date(`${isoDateStr}T00:00:00.000Z`),
+          $lt: new Date(`${isoDateStr}T23:59:59.999Z`),
+        };
+      }
     }
 
     if (!current) current = 1;
     if (!pageSize) pageSize = 10;
 
-    // Tính tổng số lượng
-    const totalItems = await this.viewingHistoryModel.countDocuments(
-      filter.query,
+    let result = await this.viewingHistoryModel
+      .find({
+        ...filter.query,
+        isDelete: { $ne: true },
+      })
+      .populate<{ videoId: { videoDescription: string } }>('videoId')
+      .sort({ updatedAt: -1, ...sort });
+
+    if (searchValue) {
+      result = result.filter(
+        (item) =>
+          item.videoId &&
+          item.videoId.videoDescription &&
+          item.videoId.videoDescription
+            .toLowerCase()
+            .includes(searchValue.toLowerCase()),
+      );
+    }
+
+    const paginatedResult = result.slice(
+      (current - 1) * pageSize,
+      current * pageSize,
     );
-    // Tính tổng số trang
-    const totalPages = Math.ceil(totalItems / pageSize);
-    const skip = (current - 1) * pageSize;
-    const result = await this.viewingHistoryModel
-      .find(filter.query) // filter đã bao gồm userId
-      .limit(pageSize)
-      .skip(skip)
-      .populate('videoId') // Lấy thông tin chi tiết của video
-      .sort({ createdAt: -1, ...sort });
-    return {
-      meta: {
-        current, // trang hiện tại
-        pageSize, // số lượng bản ghi
-        pages: totalPages, // tổng số trang với điều kiện query
-        total: totalItems, // tổng số bản ghi
-      },
-      result,
-    };
+
+    return { result: paginatedResult };
   }
 
   findAll() {
@@ -86,5 +119,21 @@ export class ViewinghistoryService {
   }
   async findByVideoId(videoId: string, userId: string) {
     return await this.viewingHistoryModel.findOne({ videoId, userId });
+  }
+  async clearAll(clearAllViewingHistoryDto: ClearAllViewingHistoryDto) {
+    const userId = clearAllViewingHistoryDto.userId;
+    return await this.viewingHistoryModel.updateMany(
+      { userId },
+      { isDelete: true },
+      { new: true },
+    );
+  }
+  async clearOne(clearOneViewingHistoryDto: ClearOneViewingHistoryDto) {
+    const id = clearOneViewingHistoryDto.id;
+    return await this.viewingHistoryModel.findOneAndUpdate(
+      { _id: id },
+      { isDelete: true },
+      { new: true },
+    );
   }
 }
